Changing key for downloaded files to History
Since the data type changed in the user defaults, it's better to change
the key.

Also fixed a typo.

NOTE: If you happened to have built from the previous commit, you'll
need to delete your preferences, or at least the downloadedFiles key.
Putting in a special check there, just for me seems like a waste of
code.